Added basic system tray icon and associated functionality
Ignore-this: 74e7ee5d0c5950dfd695dc8101461a17

Application is not closed when the main window's X button is clicked, it's hidden instead, it can be displayed
again by clicking in the system tray icon or choosing 'Show' from the context menu.

This functionality is only available if the system tray is available (as reported by Qt) and if
the desktop environment is not Unity, since the system tray replacement it ships with doesn't work.

darcs-hash:20130703145245-56b4d-af6822b4eee8a8b1e187f8e371aad6e4b1006676